created reusable code
