+++ src/Pages/Search.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import Pagination from "../components/Pagination"
 import axios from "axios";
-import { img_500, unavailable } from "../components/Config";
+import { img_500, unavailable } from "../Services/Config";
 import { Fetching } from "./Trending";
 import { lazy } from 'react';
 const Modal = lazy(() => import('../components/Modal'));
