+++ src/Services/navLinksData.ts
@@ -1,0 +1,23 @@
+export const data = [
+    {
+        id: '1',
+        name: "Home",
+        link: '/',
+    },
+    {
+        id: '2',
+        name: "Movies",
+        link: '/movies',
+    },
+    {
+        id: '3',
+        name: "TV Series",
+        link: '/tv',
+    },
+    {
+        id: '5',
+        name: "Trending",
+        link: '/trending',
+    },
+    {id:'6', name: 'Watch later', link: '/watchlater'}
+]
