+++ src/components/MobileNavbar.tsx
@@ -4,32 +4,10 @@ import { useSelector } from 'react-redux';
 import { RootState } from '../store';
 import { IoClose } from 'react-icons/io5';
 import { FaBars } from "react-icons/fa6";
+import { data } from '../Services/navLinksData';
 const Navbar = () => {
     const [isOpen, setIsOpen] = useState(false);
     const cartProducts = useSelector((state:RootState)=> state.watchlater)
-    const data = [
-        {
-            id: '1',
-            name: 'Home',
-            link: '/',
-        },
-        {
-            id: '2',
-            name: 'Movies',
-            link: '/movies',
-        },
-        {
-            id: '3',
-            name: 'TV Series',
-            link: '/tv',
-        },
-        {
-            id: '5',
-            name: 'Trending',
-            link: '/trending',
-        },
-        {id:'6', name: 'Watch Later', link: '/watchlater'}
-    ];
 
     const handleClick = () => {
         setIsOpen(!isOpen);
@@ -37,7 +15,7 @@ const Navbar = () => {
 
     return (
         <>
-            <div className='position-relative'>
+            <nav className='position-relative'>
                 <span className='bbb' onClick={handleClick}>
                     <FaBars size="35px" className='bbb-icon'/>
                 </span>
@@ -68,7 +46,7 @@ const Navbar = () => {
                         </ul>
                     </nav>
                 </div>
-            </div>
+            </nav>
         </>
     );
 };
