+++ src/components/Navbar.tsx
@@ -1,32 +1,11 @@
 import { useSelector } from "react-redux";
 import { NavLink } from "react-router-dom";
 import { RootState } from "../store";
+import { data } from "../Services/navLinksData";
 
-const Footer = () => {
+const Navbar = () => {
     const cartProducts = useSelector((state:RootState)=> state.watchlater)
-    const data = [
-        {
-            id: '1',
-            name: "Home",
-            link: '/',
-        },
-        {
-            id: '2',
-            name: "Movies",
-            link: '/movies',
-        },
-        {
-            id: '3',
-            name: "TV Series",
-            link: '/tv',
-        },
-        {
-            id: '5',
-            name: "Trending",
-            link: '/trending',
-        },
-        {id:'6', name: 'Watch later', link: '/watchlater'}
-    ]
+    
    
   return (
     <>
@@ -48,4 +27,4 @@ const Footer = () => {
   )
 }
 
-export default Footer;
+export default Navbar;
